Fix flow control, ftou and D3D10 disassembly gaps
A run of defects found by sweeping, mostly of one kind: something that
failed quietly rather than loudly, with the wrong answer recorded in a
fixture as the expected one.

An if with an empty branch. fxc writes a continue as an if with nothing in
its true branch, and EndIf reads TrueBody.Last().Outputs in four places to
build its phi nodes. A branch carrying no outputs is the case the
surrounding code already handles for a register assigned on only one side,
so an empty dictionary routes it there. Found by sweeping flow control
across ps_3_0 and ps_4_0: nested breaks, nested continues, return inside a
loop, do-while, switch fallthrough, switch inside a loop.

ftou, which existed in the opcode enum and one switch and nowhere else -
not in the disassembler, the operand tables, the integer analysis or the
parser. Found by recompiling ShaderSources/*_instruction, a corpus that
had not been used: it is register level HLSL, so it compiles back into
bytecode with far more temp traffic than the AST fixtures.

Its fixture then would not recompile, which turned up something bigger.
WriteTemporaryVariableDeclarations always wrote float and never asked the
integer analysis, so `r0.y << 1` came out shifting a float. A register is
now declared int when every written component is one the analysis never
saw a float in; one carrying both stays float. Four instruction outputs go
from not compiling to compiling.

They had been broken unnoticed because RecompileTests only ever compiled
the AST writer, so InstructionOutputCompiles now does the same for
HlslSimpleWriter, with its own known failure list - the two writers fail
on different things. 166 new cases, nine failing, each recorded with what
is wrong rather than that it is wrong. Both tests drew their cases from
Shaders(), which sets the display name, so a failure pointed at whichever
one you assumed; they name themselves apart now.

Twelve D3D10 opcodes disassembled to their enum name with no operands,
each recorded in a .asm fixture as the expected answer - sample_cmp
disassembled to `SampleCLZ`. AsmWriter has two opcode switches and only
the D3D9 default had been made to throw. Checked against fxc -dumpbin:
dcl_input_sgv, dcl_input_ps_sgv and retc_nz match exactly, sample_c_lz
differs only by this project truncating source swizzles to the destination
mask as it does everywhere else.

Two defects are recorded rather than fixed. ps_3_0/component_chain writes
components of one register inside a loop, each reading what the previous
wrote; Outputs is keyed by register component and keeps only the last, so
the temp variables the earlier ones defined are printed but never
declared. logical_and and saturate_step apply a bitwise operator to the
mask a float comparison writes; counting comparisons as integer-producing
types the register correctly, makes logical_and compile, passes every
recompile test, and turns saturate_step's `? 1 : 0` into `? 1065353216 : 0`
because that mask is anded with the bits of 1.0f. Compiling and wrong is
worse, so it stays reverted and the entry says so.

+++ HlslDecompiler.Tests/ShaderSources/cs_4_1_instruction/compute_loop.fx
@@ -4,7 +4,7 @@ RWStructuredBuffer<float4> Out : register(u0);
 [numthreads(64, 1, 1)]
 void main(uint3 sv_dispatchthreadid : SV_DispatchThreadID)
 {
-	float3 r0;
+	int3 r0;
 	float4 r1;
 	float4 r2;
 	r0.x = sv_dispatchthreadid.x << 2;
