+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/int_arithmetic.fx
@@ -4,7 +4,7 @@ float4 main(float4 texcoord : TEXCOORD) : SV_Target
 {
 	float4 o;
 
-	float r0;
+	int r0;
 	r0 = 3 * n + -7;
 	r0 = r0.x;
 	o = r0.x * texcoord;
