+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/uint_loop.fx
@@ -5,7 +5,7 @@ float4 main(float4 texcoord : TEXCOORD) : SV_Target
 	float4 o;
 
 	float4 r0;
-	float2 r1;
+	int2 r1;
 	r0 = float4(0, 0, 0, 0);
 	r1.x = 0;
 	while (true) {
