+++ HlslDecompiler.Tests/ShaderSources/vs_4_0_instruction/bitwise.fx
@@ -4,7 +4,7 @@ float4 main(uint sv_vertexid : SV_VertexID) : SV_Position
 {
 	float4 o;
 
-	float2 r0;
+	int2 r0;
 	r0.x = sv_vertexid.x & 7;
 	r0.y = mask ^ 3;
 	r0.x = r0.y | r0.x;
